Rollback of changes related to java7. Should work on java6.

+++ infra/prism/src/main/java/com/evolveum/midpoint/prism/schema/DomToSchemaProcessor.java
@@ -481,8 +481,10 @@ class DomToSchemaProcessor {
 				itemDef.setMaxOccurs(-1);
 			}
 		} else {
-			itemDef.setMinOccurs(particle.getMinOccurs().intValue());
-			itemDef.setMaxOccurs(particle.getMaxOccurs().intValue());
+            itemDef.setMinOccurs(particle.getMinOccurs());
+            itemDef.setMaxOccurs(particle.getMaxOccurs());
+			//itemDef.setMinOccurs(particle.getMinOccurs().intValue());
+			//itemDef.setMaxOccurs(particle.getMaxOccurs().intValue());
 		}
 	}
